restructure signup page to be more modular
